requirements.txt
pydantic
uvicorn
tortoise-orm
git+https://github.com/nexios-labs/nexios.git@internals/eco